+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed the function of saving or not saving data in the database
 * Resolve not find name
+* Removed DEL_TEMP_INSTANCES as it is not being used
 
 # 2.0.1-beta (2024-07-17 17:01)